Fix tool_use README accuracy and rename code_rl tools.py
README fixes:
- Change ToolInterface → Tool protocol (accurate to implementation)
- Fix example return types: str → ToolResult
- Add simple_tool_result usage in examples
- Import simple_tool_result in quick example

Rename:
- code_rl/tools.py → code_rl/deepcoder_tool.py (more descriptive)
- Update import in code_env.py
